telegram-bot/requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
pyyaml==6.0.1
